+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Include Account ID in the s3bucket for access logs. It is a breaking change, that will put access logs to a new s3 bucket.
 - Change AWS CNI and AWS CNI k8s plugin log verbosity to `INFO`.
 - Change AWS CNI log file to `stdout`.